requirements.txt
openai>=1.0.0
google-api-python-client>=2.0.0
google-auth>=2.0.0
python-dotenv>=1.0.0
requests>=2.28.0
typing-extensions>=4.0.0
beautifulsoup4>=4.9.0
